scopesearch/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.1.1
httpx==0.27.2
dnspython==2.7.0
pyyaml==6.0.2
cryptography==43.0.1
